some updates based on comment
