+++ packages/codegen/src/__future__/definePlugin.test.ts
@@ -1,0 +1,13 @@
+import {describe, it, expect} from "vitest";
+import {
+  defineOazapftsPlugin,
+} from "./definePlugin";
+
+describe("defineOazapftsPlugin", () => {
+  it("should return an object", () => {
+    const result = defineOazapftsPlugin({
+      name: "test",
+    });
+    expect(result).toBeInstanceOf(Object);
+  });
+});
