+++ packages/codegen/src/__future__/definePlugin.ts
@@ -3,13 +3,12 @@ import { type SchemaObject } from "../generate";
 import ts from "typescript";
 import { Opts } from "../index";
 import minimist from "minimist";
-import assert from "node:assert";
 
-export type OazapftsCodegenPluginDefinition = {
+export type OazapftsCodegenPlugin = {
   /**
-   * Required identifier for this plugin
+   * identifier for this plugin
    */
-  name: string;
+  name?: string;
   /**
    * Additional command line arguments that can be specified
    * This will be parsed after all plugins have been loaded.
@@ -26,10 +25,10 @@ export type OazapftsCodegenPluginDefinition = {
     description: string;
     /**
      * Validate if the options you have are valid or not.
-     * options will be considered valid by default if not validator plugin is provided.
+     * should throw an error if the options are not valid.
      * @param opts - the parsed options object.
      */
-    validator?(opts: Opts): Boolean;
+    validator?(opts: Opts): void;
   };
 
   /**
@@ -162,50 +161,14 @@ export type OazapftsCodegenPluginDefinition = {
   ): void | Promise<void>;
 };
 
-export const IS_OAZAPFTS_PLUGIN = Symbol("OazapftsPlugin");
-
-export type OazapftsCodegenPlugin = OazapftsCodegenPluginDefinition & {
-  [IS_OAZAPFTS_PLUGIN]: boolean;
-};
-
-/**
- * Assigns a hidden key we can use to make sure the default export is in fact a plugin when we import it and prevent some
- * potential headaches.
- * @param plugin
- * @param value
- */
-function convertToPlugin(
-  plugin: OazapftsCodegenPluginDefinition,
-  value: boolean,
-): asserts plugin is OazapftsCodegenPlugin {
-  Object.defineProperty(plugin, IS_OAZAPFTS_PLUGIN, {
-    value,
-    enumerable: false,
-    writable: false,
-    configurable: false,
-  });
-}
 
 export function defineOazapftsPlugin(
-  plugin: OazapftsCodegenPluginDefinition,
+  plugin: OazapftsCodegenPlugin,
 ): OazapftsCodegenPlugin {
   // Do some checks to ensure that we have a valid plugin (if there are required params)
-  let isPlugin = true;
-  if (!plugin.name || plugin.name === "") {
-    console.warn("defineOazapftsPlugin: 'name' property is required!");
-    isPlugin = false;
+  if(typeof plugin !== "object" || Array.isArray(plugin) || plugin === null ){
+    throw new Error("Plugin is not an object.")
   }
-  convertToPlugin(plugin, isPlugin);
   // Maybe we do the registration right here? force module developers to use this function?
   return plugin;
-}
-
-export function isOazapftsPlugin(
-  plugin: unknown,
-): plugin is OazapftsCodegenPlugin {
-  return (
-    plugin instanceof Object &&
-    plugin.hasOwnProperty(IS_OAZAPFTS_PLUGIN) &&
-    (plugin as OazapftsCodegenPlugin)[IS_OAZAPFTS_PLUGIN]
-  );
 }
